pyproject.toml
[project]
name = "a-share-short"
version = "0.1.0"
description = "A股超短线复盘与情绪分析系统"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
pdf = [
    "reportlab>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
